fix issue #783

+++ tests/decompress-partial.c
@@ -28,7 +28,7 @@ const char source[] =
 
 int main(void)
 {
-  size_t srcLen = strlen(source);
+  int srcLen = (int)strlen(source);
   char cmpBuffer[BUFFER_SIZE];
   char outBuffer[BUFFER_SIZE];
   int cmpSize, result;
@@ -37,8 +37,8 @@ int main(void)
   cmpSize = LZ4_compress_default(source, cmpBuffer, srcLen, BUFFER_SIZE);
 
   for (i = cmpSize; i < cmpSize + 10; ++i) {
-    result = LZ4_decompress_safe_partial(cmpBuffer, outBuffer, i, (int)srcLen, BUFFER_SIZE);
-    if ((result < 0) || ((unsigned int)result != srcLen) || memcmp(source, outBuffer, srcLen)) {
+    result = LZ4_decompress_safe_partial(cmpBuffer, outBuffer, i, srcLen, BUFFER_SIZE);
+    if ((result < 0) || (result != srcLen) || memcmp(source, outBuffer, srcLen)) {
       printf("test decompress-partial error \n");
       return -1;
     }
